otica-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-jose[cryptography]>=3.3.0
httpx>=0.27.2
python-dotenv>=1.0.1
cryptography>=43.0.1
email-validator>=2.0.0

